+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 - Power VS Workspace connections are no longer removed from Transit Gateways when updating the Workspace's region
+- Users can now create Secrets Groups from the Secrets Manager page
+- Users can now add authorizations for Kubernetes and CIS to Secrets Manager from the Secrets Manager Page
 
 ### Fixes